pyproject.toml
[project]
name = "resume-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.58.2",
    "dspy-ai>=2.6.27",
    "gradio>=5.38.0",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
